Refactor filenames and move some code around
