+++ test/test_Ineasibility.jl
@@ -3,7 +3,7 @@
 # Use of this source code is governed by an MIT-style license that can be found
 # in the LICENSE.md file or at https://opensource.org/licenses/MIT.
 
-module TestInfeasibilityChecker
+module TestInfeasibility
 
 import ModelAnalyzer
 using Test
@@ -537,6 +537,6 @@ function test_iis_spare()
     return
 end
 
-end # module
+end  # module TestInfeasibility
 
-TestInfeasibilityChecker.runtests()
+TestInfeasibility.runtests()
